requirements.txt
# Core ML/AI (GPU optimized)
torch>=2.1.0+cu118
torchvision>=0.16.0+cu118
torchaudio>=2.1.0+cu118
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# GPU acceleration
cupy-cuda11x>=12.0.0  # Optional: for faster GPU operations

# RL/IL
gymnasium>=0.29.0
stable-baselines3>=2.0.0
tensorboard>=2.13.0
wandb>=0.15.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Data Processing
pandas>=2.0.0
h5py>=3.9.0
zarr>=2.16.0

# Configuration & CLI
pyyaml>=6.0
click>=8.1.0
tqdm>=4.65.0
rich>=13.0.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
ruff>=0.0.280
mypy>=1.5.0

# Video & Media
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Utilities
requests>=2.31.0
urllib3>=2.0.0
psutil>=5.9.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Windows-specific
pywin32>=306; sys_platform == "win32"
